Develop 27tech

+++ tatau_core/tatau/models.py
@@ -110,7 +110,7 @@ class TaskDeclaration(models.Model):
 
     state = fields.CharField(initial=State.DEPLOYMENT)
     current_epoch = fields.IntegerField(initial=0)
-    progress = fields.IntegerField(initial=0)
+    progress = fields.FloatField(initial=0.0)
     tflops = fields.FloatField(initial=0.0)
     results = fields.EncryptedJsonField(initial=[])
 
@@ -272,7 +272,7 @@ class TaskAssignment(models.Model):
     train_data = fields.EncryptedJsonField(required=False)
     current_epoch = fields.IntegerField(initial=0)
 
-    progress = fields.IntegerField(initial=0)
+    progress = fields.FloatField(initial=0.0)
     tflops = fields.FloatField(initial=0.0)
 
     result = fields.EncryptedCharField(required=False)
@@ -311,7 +311,7 @@ class VerificationAssignment(models.Model):
     state = fields.CharField(initial=State.INITIAL)
     train_results = fields.EncryptedJsonField(required=False)
 
-    progress = fields.IntegerField(initial=0)
+    progress = fields.FloatField(initial=0.0)
     tflops = fields.FloatField(initial=0.0)
     result = fields.EncryptedJsonField(required=False)
     error = fields.EncryptedCharField(required=False)
